clean ups - coding standard

+++ plugins/hifiKinect/src/KinectPlugin.h
@@ -16,12 +16,6 @@
 #define WIN32_LEAN_AND_MEAN // Exclude rarely-used stuff from Windows headers
 #endif
 
-// glm files
-
-#include <glm/glm.hpp>
-#include <glm/gtc/quaternion.hpp>
-
-
 // Windows Header Files
 #include <windows.h>
 
@@ -65,12 +59,12 @@ public:
     virtual void loadSettings() override;
 
 private:
-	// add variables for moving average
+    // add variables for moving average
 
-	glm::quat _q_barL;
-	glm::quat _q_barR;
-    float _deltaL = (float)0.5;
-	float _deltaR = (float)0.5;
+    glm::quat _q_barL;
+    glm::quat _q_barR;
+    float _deltaL = 0.5f;
+    float _deltaR = 0.5f;
 
 protected:
